==> Ranalyses/01_NJ_tree.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---

# NJ/Phylo Analyses
```{r, include=FALSE}
knitr::opts_chunk$set(echo = T, warning = F, message = F, eval = T, results = 'hide',
                      fig.align = 'center', fig.width = 8, fig.height = 8)
knitr::opts_knit$set(root.dir = here::here())
```


```{r}
library(tidyverse)
library(vcfR)
library(vcfRmanip)
library(ape)
library(adegenet)
library(poppr)
library(ggtree)
```

```{r}
#---------------------------------------------------- 
# Read In
#---------------------------------------------------- 
mtdt <- read_tsv("mtdt/mtdt_clean_sm.tab.txt")
vcfbinomiss <- readRDS("data/salmonellavcf/consensus_thinned_passed.joint.vcf.gz")



#............................................................
# distance matrix large
#...........................................................
geneDist <- poppr::nei.dist( vcfR::vcfR2genind(vcfbinomiss))
saveRDS(geneDist, "data/salmonellavcf/geneDist.RDS")
myTree <- nj(as.matrix(geneDist))

```




## Pretty Tree
```{r}
#......................
# make prettty tree
#......................
tidytree <-  treeio::as.treedata(myTree)

# tidy and make tree
treemtdt <- mtdt %>%
  dplyr::mutate(label = biosample_acc,
                PeanutButter = pb,
                color = dplyr::case_when(
                  culprit == "y" ~ "Culprit",
                  pb == "y" ~ "Peanuts",
                  pb == "n" & culprit == "n" & biosample_acc != "parr" ~ "Seq2022",
                  biosample_acc == "parr" ~ "Parr!"
                ))

tidytree <- left_join(tidytree, treemtdt, by = 'label')

# make plot
treeplot_large <- tidytree %>% 
  ggtree(.,
         #layout="fan",
         aes(color = color)) +
  geom_tippoint(aes(color =  color, shape = color),
                show.legend = F,
                size = 1) +
  geom_text(aes(label = cntnm), size = 6, fontface = "bold", hjust = -0.15) +
  theme_tree() +
  scale_color_manual(name = "Source",
                     values = c("#FF0018", "#B22234", "#008018", "#0000F9", "#86007D")) +
  theme(legend.position = "none") +
  xlim(0, 0.15)

```

```{r}
#............................................................
# distance matrix small
#...........................................................
smsmpls <- mtdt %>% 
  dplyr::filter(cntnm %in% c("Canada_2022.4",
                             "USA:GA_2020.1",
                             "Syria_2022.6",
                             "USA:AR_2012.1",
                             "USA:NM_2011.2",
                              "USA:NM_2011.1",
                               "USA:NM_2011.11",
                              "Egypt_2022.2",
                              "Egypt_2022.1",
                              "India_2022.4",
                              "Parr!",
                              "USA: KY_2010.1",
                              "Bangladesh_2022.1",
                              "Ghana_2022.1",
                              "Mexico_2022.1"
                             )) %>% 
  dplyr::select(biosample_acc) %>% 
  dplyr::pull()

vcfsm <- vcfRmanip::select_samples(vcfbinomiss, smplvctr = smsmpls)
geneDist <- poppr::nei.dist( vcfR::vcfR2genind(vcfsm))
myTree <- nj(as.matrix(geneDist))
tidytree <-  treeio::as.treedata(myTree)
#............................................................
# small tree
#...........................................................
tidytree <- left_join(tidytree, treemtdt, by = 'label')

# make plot
treeplot_sm <- tidytree %>% 
  ggtree(.,
         #layout="fan",
         aes(color = color)) +
  geom_tippoint(aes(color =  color, shape = color),
                show.legend = F,
                size = 1) +
  geom_text(aes(label = cntnm), size = 10, fontface = "bold", hjust = -0.15) +
  theme_tree() +
  scale_color_manual(name = "Source",
                     values = c("#FF0018", "#B22234", "#008018", "#0000F9", "#86007D")) +
  theme(legend.position = "none") +
  xlim(0, 0.2)


```



```{r}
jpeg(filename = "~/Desktop/salmonella_tree.jpg", width = 30, height = 15, units = "in", res = 500)
cowplot::plot_grid(treeplot_large, treeplot_sm,
                   align = "h", rel_widths = c(0.7, 0.3),
                   labels = c("(A)", "(B)"),
                   label_size = 42)
graphics.off()


```
