==> Ranalyses/02_networks.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---
```{r, include=FALSE}
knitr::opts_chunk$set(echo = T, warning = F, message = F, eval = T, results = 'hide',
                      fig.align = 'center', fig.width = 8, fig.height = 8)
knitr::opts_knit$set(root.dir = here::here())
```

```{r}
library(tidyverse)
library(tidygraph)
library(ggraph)
library(cowplot)
```

```{r}
mtdt <- read_tsv("mtdt/mtdt_clean_sm.tab.txt") %>%
  dplyr::mutate(cntnm = ifelse(culprit == "y", "Culprit", cntnm))

vcfbinomiss <- readRDS("data/salmonellavcf/consensus_thinned_passed.joint.vcf.gz")
geneDist <- poppr::nei.dist( vcfR::vcfR2genind(vcfbinomiss))

# convert to network
gennetwork <- broom::tidy(geneDist) %>% 
  tidygraph::as_tbl_graph(., directed = F) 

# bring in meta data
gennetwork <- gennetwork %>% 
  tidygraph::activate(nodes) %>% 
  dplyr::mutate(biosample_acc = name) %>% 
  dplyr::left_join(., mtdt, by = "biosample_acc")

#............................................................
# centrality
#...........................................................

ng <- gennetwork %>% 
  tidygraph::activate(edges) %>%
  dplyr::filter(distance > 0.05) %>% 
  tidygraph::activate(nodes) %>%
  mutate(centrality = centrality_closeness()) %>% 
  as_tibble() 

summary(ng$centrality)  


netplot <- gennetwork %>% 
  tidygraph::activate(edges) %>%
  dplyr::filter(distance > 0.05) %>% 
  tidygraph::activate(nodes) %>%
  mutate(centrality = centrality_betweenness()) %>% 
  ggraph(layout = "kk") +
  geom_edge_density(aes(size = distance), alpha = 0.25) +
   geom_edge_link(alpha = 0.05, size = 0.05) +
  geom_node_point(aes(color = centrality), size = 6) + 
  scale_edge_size("Nei's D") +
  scale_color_viridis_c("Betweeness \n Centrality",) +
  geom_node_text(aes(label = cntnm), size = 7, fontface = "bold") +
  theme_graph() +
  theme(legend.position = "bottom",
        legend.title = element_text(family = "Helvetica", vjust = 0.85, hjust = 0.5, size = 13, face = "bold"),
        legend.text = element_text(family = "Helvetica", hjust = 1, size = 11, face = "bold", angle = 45))


#............................................................
# out
#...........................................................
jpeg(filename = "~/Desktop/salmonella_net.jpg", width = 30, height = 15, units = "in", res = 500)
netplot
graphics.off()



```
